+++ packages/@luban/cli-shared-utils/src/index.ts
@@ -1,7 +1,7 @@
 import { hasGit, hasProjectGit, isLinux, isMacintosh, isWindows, installedBrowsers } from "./env";
 import { IpcMessenger } from "./ipc";
 import { log, warn, error, info, done, clearConsole } from "./logger";
-import { loadModule, resolveModule, clearModule } from "./module";
+import { loadModule, resolveModule, clearModule, loadFile } from "./module";
 import { openBrowser } from "./openBrowser";
 import { Spinner } from "./spinner";
 import { createSchema, validate, validateSync } from "./validate";
@@ -24,6 +24,7 @@ export {
   loadModule,
   resolveModule,
   clearModule,
+  loadFile,
   openBrowser,
   Spinner,
   createSchema,